add safety checks to forbid players from editing characters that do not belong to them

+++ nbot.py
@@ -33,6 +33,7 @@ bot = commands.Bot(command_prefix='^', intents=intents)
 async def on_ready():
     print("The bot is alive, and the database is operational.")
     await bot.tree.sync(guild=GUILD_ID)
+
     conn.commit()
 
 @bot.event  # return if user is the one sending a message
@@ -44,19 +45,24 @@ async def on_message(message):
     try:
         tupper = Tupper(message.author.id, cursor)
         tupper.add_xp_by_bracket(word_count, tupper_try)
+
         conn.commit()
+
     except Exception as e:
         print(f"Error in on_message:\n{e}")
+        return
 
 # User Commands
 
-@bot.tree.command(name='register_me', description="Creates a character for use. Require the following arguments: name, level, starting tupper bracket.", guild=GUILD_ID)
+@bot.tree.command(name='register_me', description="Adds you to the database so you can use other functions of the bot.", guild=GUILD_ID)
 async def RegisterPlayer(interaction: discord.Interaction):
     discord_id = interaction.user.id
     player = Player(discord_id, cursor)
     try: 
         success_message = player.register_player(interaction.user.name)
+
         conn.commit()
+
         await interaction.response.send_message(success_message)
     except Exception as e: 
         await interaction.response.send_message(f"You are already in the database, {interaction.user.name}\n{e}")
@@ -71,9 +77,12 @@ async def SeeMyCharacters(interaction: discord.Interaction):
         command_embed_instance = EmbedWrapper().return_base_embed()
         command_embed_instance.add_field(name="Command Output for SeeMyCharacters", value=f"Success! \n\n {message}")
         await interaction.response.send_message(embed=command_embed_instance)
+
         conn.commit()
+
     except Exception as e:
         await interaction.response.send_message(f"Error retrieving characters:\n{e}")
+        return
 
 @bot.tree.command(name='return_players', description="Returns list of player discord_ids for debugging purposes.", guild=GUILD_ID)
 async def ReturnPlayers(interaction: discord.Interaction):
@@ -81,9 +90,12 @@ async def ReturnPlayers(interaction: discord.Interaction):
         cursor.execute("SELECT * FROM players;")
         rows = cursor.fetchall()
         await interaction.response.send_message(rows)
+
         conn.commit()
+
     except Exception as e:
         await interaction.response.send_message(f"Error retrieving players:\n{e}")
+        return
 
 @bot.tree.command(name='add_character', description="Creates a character", guild=GUILD_ID)
 async def AddCharacter(interaction: discord.Interaction, name: str):
@@ -92,12 +104,16 @@ async def AddCharacter(interaction: discord.Interaction, name: str):
     character = Character(discord_id, cursor)
     try: 
         message = character.register_character(discord_id, name)
+
         conn.commit()
+
         await interaction.response.send_message(message)
     except sqlite3.IntegrityError as e:
         await interaction.response.send_message(f"Error, character name already in use:\n{e}")
+        return
     except Exception as e:
         await interaction.response.send_message(f"Error adding character:\n{e}")
+        return
 
 @bot.tree.command(name='add_tupper', description="Adds a tupper.", guild=GUILD_ID)
 async def AddTupper(interaction: discord.Interaction, bracket: str, character_name: str):
@@ -122,28 +138,51 @@ async def AddTupper(interaction: discord.Interaction, bracket: str, character_na
             value=f"Success! \n\n **Player:** {interaction.user.name} \n\n **Tupper:** {emb_tupper_bracket} \n\n **Bound Character:** {emb_character_name}"
         )
         await interaction.response.send_message(embed=command_embed_instance)
+
         conn.commit()
+
     except Exception as e:
         await interaction.response.send_message(f"Error adding tupper:\n{e}")
+        return
 
 @bot.tree.command(name='remove_tupper', description="Removes a tupper.", guild=GUILD_ID)
 async def RemoveTupper(interaction: discord.Interaction, bracket: str):
+    discord_id = interaction.user.id
+    tupper = Tupper(discord_id, cursor)
+    if not tupper.tupper_belongs_to_player(bracket):
+        await interaction.response.send_message("This tupper does not belong to you.")
+        return
+
     valid_brackets = bracket.endswith(':')  # tupper brackets must end with a colon 
     if not valid_brackets:
         await interaction.response.send_message(f"Make sure that your tupper brackets end with a colon.")
         return
     discord_id = interaction.user.id
     try: 
-        tupper = Tupper(discord_id, cursor)
         message = tupper.delete_tupper(bracket)
         await interaction.response.send_message(message)
+
         conn.commit()
+
     except Exception as e: 
         await interaction.response.send_message(f"Error removing tupper:\n{e}")
+        return
 
 @bot.tree.command(name='remove_xp_from_character', description="Command that removes xp by discord_id,", guild=GUILD_ID)
 async def RemoveXPFromCharacter(interaction: discord.Interaction, character_name: str, xp: str):
+
+    if int(xp) != abs(int(xp)): 
+        await interaction.response.send_message("Nice try. Make it positive.")
+        return
+
     discord_id = interaction.user.id
+    character = Character(discord_id, cursor)
+    character_id = character.get_owned_character(character_name)
+    if not character_id:
+        await interaction.response.send_message("You do not own this character.")
+        return
+    # proceed safely with character_id
+
     try:  # see if can turn string xp into int xp, if not, return helpful error message
         xp_int = int(xp)
     except ValueError as e: 
@@ -153,13 +192,23 @@ async def RemoveXPFromCharacter(interaction: discord.Interaction, character_name
         character = Character(discord_id, cursor)
         message = character.remove_xp(character_name, xp_int, discord_id)
         await interaction.response.send_message(message)
+
         conn.commit()
+
     except Exception as e: 
         await interaction.response.send_message(f"Error removing XP:\n{e}")
+        return
 
 @bot.tree.command(name='set_level_of_character', description="Command that removes xp by discord_id,", guild=GUILD_ID)
 async def SetCharacterLevel(interaction: discord.Interaction, character_name: str, level: str):
     discord_id = interaction.user.id
+    character = Character(discord_id, cursor)
+    character_id = character.get_owned_character(character_name)
+    if not character_id:
+        await interaction.response.send_message("You do not own this character.")
+        return
+    # proceed safely with character_id
+
     try:  # see if can turn string xp into int xp, if not, return helpful error message
         level_int = int(level)
         if not 1 <= level_int <= 20:
@@ -171,10 +220,53 @@ async def SetCharacterLevel(interaction: discord.Interaction, character_name: st
     try:
         character = Character(discord_id, cursor)
         character.set_level(character_name, level_int, discord_id)
+
         conn.commit()
+
         await interaction.response.send_message(f"Character {character_name} level set to {level_int}.")
     except Exception as e: 
         await interaction.response.send_message(f"Error setting level:\n{e}")
+        return
+
+@bot.tree.command(name='delete_character', description="Command that deleted a character.", guild=GUILD_ID)
+async def DeleteCharacter(interaction: discord.Interaction, character_name: str):    
+    discord_id = interaction.user.id
+    character = Character(discord_id, cursor)
+    character_id = character.get_owned_character(character_name)
+    if not character_id:
+        await interaction.response.send_message("You do not own this character.")
+        return
+    # proceed safely with character_id
+    try: 
+        character = Character(discord_id, cursor)
+        message = character.delete_character(character_name, discord_id)
+
+        conn.commit()
+
+        await interaction.response.send_message(message)
+    except Exception as e: 
+        await interaction.response.send_message(f"Character deletion has failed: \n {e}.")
+        return
+    
+@bot.tree.command(name='rename_character', description="Command that renames a pre-existing character.", guild=GUILD_ID)
+async def RenameCharacter(interaction: discord.Interaction, character_name: str, new_character_name: str):    
+    discord_id = interaction.user.id
+    character = Character(discord_id, cursor)
+    character_id = character.get_owned_character(character_name)
+    if not character_id:
+        await interaction.response.send_message("You do not own this character.")
+        return
+    # proceed safely with character_id
+    try: 
+        character = Character(discord_id, cursor)
+        message = character.rename_character(character_name, new_character_name)
+
+        conn.commit()
+
+        await interaction.response.send_message(message)
+    except Exception as e: 
+        await interaction.response.send_message(f"Character renaming has failed: \n {e}.")
+        return
 
 @bot.event  # close database connection if bot goes offline
 async def shutdown():
